refactor: Hands 객체를 삭제하고 해당 구현을 Gambler로 이동
